Refactor code: improve logic and code structure
- Refactored the logic that updates the `offset` for the previous and
next buttons, as well as the count for `currentPage` and `totalPages`.
My previous implementation was causing several bugs when trying to
implement localStorage. This approach prevents those bugs and also
makes the code more readable.
Additionally, the buttons `previous` and `next` will be disabled when
the offset is invalid.

- The `updatePageCount` function was split into two different functions:
`updateTotalPages` and `updateCurrentPage` to separate concerns.

- Moved the `displayPokemonList` function to the `ui.js` file. This is
a more appropriate file for this functionality.

- Removed the default parameter in `getPokemonList()`. The default
parameter was also causing a bug with my current implementation of
pagination where the page wouldn't go back to 1 (offset 0) when using
the previous button. Invoking `displayPokemonList(0)` solved the issue.