Refactor list elements to remove unnecessary code
- Moved the data attr "pokemon-number" from the element "pokemon-img"
to the element "pokemon-box" to improve the `setupPokemonModalButton`
function. This change makes the function straightforward, improving
readability.

- Renamed the "$pokemonEl" variable to "$pokemonBox" for consistency
with other parts of the code. Additionally, I removed the Bootstap class
"justify-content-evenly" and added it through the CSS file instead. The
result is the same, and it avoids an unnecessary long line of code.