Refactorize codebase to implement try...catch
- Updated the functions that interact with the localStorage in order to
implement try...catch when displaying pages or modals. The functions
responsible for retrieving data will throw an error if there's no data
to retrieve, essentially exiting the "try" statement and retrieving the
data directly from the API.

- Renamed the files `api.js` and `localStorage.js` to `pokemon.js`.
These names are more consistent throughout the application and align
better with the overall naming of it.

- Renamed the imported functions to make them more self-explanatory.
For example: "getPokemonList as getPokemonListFromApi", "getPokemon as
getPokemonFromStorage", etc.

- Renamed the variable "pokemonPerPage" to "POKEMON_PER_PAGE" and moved
it to `api/pokemon.js` which is more appropriate for its purpose.